refactor: merge in the nav plugin to core
