+++ packages/kickflip-cli/app/Listeners/SetupSiteNav.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Kickflip\RouterNavPlugin\Listeners;
+namespace Kickflip\Listeners;
 
 use Illuminate\Config\Repository;
 use Illuminate\Support\Facades\View;
