--- packages/kickflip-router-nav-plugin/.github/workflows/auto_closer.yaml
@@ -1,14 +1,0 @@
-name: Auto Closer PR
-
-on:
-  pull_request_target:
-    types: [opened]
-
-jobs:
-  run:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: superbrothers/close-pull-request@v3
-      with:
-        # Optional. Post a issue comment just before closing a pull request.
-        comment: "This repository's content is generated by mallardduck/kickflip-monorepo. Please open pull request to [mallardduck/kickflip-monorepo](https://github.com/mallardduck/kickflip-monorepo)."
