--- packages/kickflip-router-nav-plugin/README.md
@@ -1,17 +1,0 @@
-# Kickflip CLI Router Nav Plugin
-#### A plugin to provide Kickflip with pseudo-router functionality.
-
-## A message to Russian 🇷🇺 people
-
-If you currently live in Russia, please read [this message](./ToRussianPeople.md).
-
-
-------
-
-## Contributing
-
-Please direct all PRs and Issues to the main [Kickflip Repo](https://github.com/KickflipCli/kickflip-src).
-
-## License
-
-Kickflip CLI is an open-source software licensed under the MIT [license](LICENSE.md).
