--- packages/kickflip-router-nav-plugin/composer.json
@@ -1,35 +1,0 @@
-{
-    "name": "kickflip/kickflip-router-nav-plugin",
-    "description": "A Kickflip CLI plugin which provides  pseudo-router functionality with helpful URL and Nav generation features.",
-    "keywords": ["kickflip plugin", "kickflip cli plugin", "static site generator plugin"],
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Dan Pock",
-            "email": "self@danpock.me"
-        }
-    ],
-    "require": {
-        "php": "^8.0",
-        "kickflip/kickflip-cli": "^0.10",
-        "illuminate/routing": "^9.0",
-        "illuminate/support": "^9.0"
-    },
-    "config": {
-        "optimize-autoloader": true,
-        "preferred-install": "dist",
-        "sort-packages": true
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-main": "0.10-dev"
-        }
-    },
-    "autoload": {
-        "psr-4": {
-            "Kickflip\\RouterNavPlugin\\": "src/"
-        }
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true
-}
