+++ phpunit.xml.dist
@@ -18,9 +18,6 @@
         <testsuite name="Features">
             <directory suffix="Test.php">tests/Feature</directory>
         </testsuite>
-        <testsuite name="PluginsTests">
-            <directory suffix="Test.php">tests/Plugins</directory>
-        </testsuite>
         <testsuite name="DocsSiteTests">
             <directory suffix="Test.php">tests/DocsSite</directory>
         </testsuite>
